refactor: remove Contentful-era display Drink type
Unify around the DB schema Drink type. The old display type had false
nullability left over from Contentful. Centralizes DB-to-display mapping
in withPlaceholderImages instead of duplicating it across 4 loaders.
