+++ app/routes/_app.search/route.tsx
@@ -4,7 +4,8 @@ import { defaultPageDescription, defaultPageTitle } from '#/app/core/config';
 import { getEnvVars } from '#/app/utils/env.server';
 import { withPlaceholderImages } from '#/app/utils/placeholder-images.server';
 import { DrinkList } from '#/app/drinks/drink-list';
-import type { AppRouteHandle, Drink } from '#/app/types';
+import type { Drink } from '#/app/db/schema';
+import type { AppRouteHandle } from '#/app/types';
 import { NoDrinksFound } from './no-drinks-found';
 import { NoSearchTerm } from './no-search-term';
 import { SearchForm } from './search-form';
